refactor: move swr to frontend page
