+++ src/application/frontend/swr/useExerciseTechniques.ts
