+++ src/application/frontend/swr/usePlanningTypes.ts
